ftCo_0A01 matches (#1869)

+++ src/melee/ft/ft_0877.c
@@ -132,7 +132,7 @@ bool ft_80087A18(HSD_GObj* gobj)
     return false;
 }
 
-s8 ft_80087A80(HSD_GObj* gobj)
+int ft_80087A80(HSD_GObj* gobj)
 {
     Fighter* fp = GET_FIGHTER(gobj);
     return fp->smash_attrs.x2134_vibrateFrame;
